backend/requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1

# Vector Database
chromadb==0.4.18

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
pandas==2.1.4

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Environment Management
python-dotenv==1.0.0

# Evaluation and Monitoring
langsmith==0.0.69

# Text Processing
tiktoken==0.5.2
sentence-transformers==2.2.2

# Utilities
requests==2.31.0
pydantic==2.5.2
numpy==1.24.3

# Development
python-json-logger==2.0.7
